Working on Upgrade UI.
Added UI for the upgrades, displaying what you have and how many you have. Currently on have it done for 1 upgrade, but can easily expand it to more.
Added tooltip system so you can see what your upgrades do. (Pulled this from a previous project of mine)
Imported LeanTween to assist with tooltip system, which is why the changed file size is so large.